requirements.txt
streamlit
langchain>=0.0.200
openai
python-dotenv